Moved flatsat test scripts out of flatsat_tests folder, renamed them to include flatsat in their names, fixed comments, and edited how other files get imported into each test script
